Point refactoring complete
